updates to get unit tests passing following main merge
