+++ product/consul.go
@@ -2,18 +2,19 @@ package product
 
 import (
 	"context"
+	"fmt"
 	"path/filepath"
 
 	"github.com/hashicorp/hcdiag/hcl"
 	"github.com/hashicorp/hcdiag/redact"
 	"github.com/hashicorp/hcdiag/runner/debug"
 	"github.com/hashicorp/hcdiag/runner/do"
+	logs "github.com/hashicorp/hcdiag/runner/log"
 
 	"github.com/hashicorp/go-hclog"
 
 	"github.com/hashicorp/hcdiag/client"
 	"github.com/hashicorp/hcdiag/runner"
-	logs "github.com/hashicorp/hcdiag/runner/log"
 )
 
 const (
@@ -76,6 +77,21 @@ func NewConsulWithContext(ctx context.Context, logger hclog.Logger, cfg Config) 
 
 // consulRunners generates a slice of runners to inspect consul.
 func consulRunners(ctx context.Context, cfg Config, api *client.APIClient, l hclog.Logger) ([]runner.Runner, error) {
+	var r []runner.Runner
+
+	// Set up Command runners
+	for _, cc := range []runner.CommandConfig{
+		{Command: "consul version", Redactions: cfg.Redactions},
+		{Command: fmt.Sprintf("consul debug -output=%s/ConsulDebug -duration=%s -interval=%s", cfg.TmpDir, cfg.DebugDuration, cfg.DebugInterval), Redactions: cfg.Redactions},
+		{Command: "consul operator raft list-peers -stale=true", Redactions: cfg.Redactions},
+	} {
+		c, err := runner.NewCommandWithContext(ctx, cc)
+		if err != nil {
+			return nil, err
+		}
+		r = append(r, c)
+	}
+
 	dbg, err := debug.NewConsulDebug(
 		debug.ConsulDebugConfig{
 			Redactions: cfg.Redactions,
@@ -87,27 +103,30 @@ func consulRunners(ctx context.Context, cfg Config, api *client.APIClient, l hcl
 	if err != nil {
 		return nil, err
 	}
+	r = append(r, dbg)
 
-	r := []runner.Runner{
-		runner.NewCommand("consul version", "string", cfg.Redactions),
+	// Set up HTTP runners
+	for _, hc := range []runner.HttpConfig{
+		{Client: api, Path: "/v1/agent/self", Redactions: cfg.Redactions},
+		{Client: api, Path: "/v1/agent/metrics", Redactions: cfg.Redactions},
+		{Client: api, Path: "/v1/catalog/datacenters", Redactions: cfg.Redactions},
+		{Client: api, Path: "/v1/catalog/services", Redactions: cfg.Redactions},
+		{Client: api, Path: "/v1/namespace", Redactions: cfg.Redactions},
+		{Client: api, Path: "/v1/status/leader", Redactions: cfg.Redactions},
+		{Client: api, Path: "/v1/status/peers", Redactions: cfg.Redactions},
+		{Client: api, Path: "/v1/agent/members?cached", Redactions: cfg.Redactions},
+	} {
+		c, err := runner.NewHTTPWithContext(ctx, hc)
+		if err != nil {
+			return nil, err
+		}
+		r = append(r, c)
+	}
 
-		// TODO(dcohen) ConsulDebug -- adjust during merge
-		dbg,
-
-		runner.NewCommand("consul operator raft list-peers -stale=true", "string", cfg.Redactions),
-
-		runner.NewHTTP(api, "/v1/agent/self", cfg.Redactions),
-		runner.NewHTTP(api, "/v1/agent/metrics", cfg.Redactions),
-		runner.NewHTTP(api, "/v1/catalog/datacenters", cfg.Redactions),
-		runner.NewHTTP(api, "/v1/catalog/services", cfg.Redactions),
-		runner.NewHTTP(api, "/v1/namespace", cfg.Redactions),
-		runner.NewHTTP(api, "/v1/status/leader", cfg.Redactions),
-		runner.NewHTTP(api, "/v1/status/peers", cfg.Redactions),
-		runner.NewHTTP(api, "/v1/agent/members?cached", cfg.Redactions),
-
+	r = append(r,
 		logs.NewDocker("consul", cfg.TmpDir, cfg.Since, cfg.Redactions),
 		logs.NewJournald("consul", cfg.TmpDir, cfg.Since, cfg.Until, cfg.Redactions),
-	}
+	)
 
 	// try to detect log location to copy
 	if logPath, err := client.GetConsulLogPath(api); err == nil {
